refactor: separated utils
