+++ src/app/(store)/list/[[...params]]/page.tsx
@@ -14,26 +14,13 @@ import DropDownList from "@/components/UI/dropDown";
 import LineList from "@/components/UI/lineList";
 import { SK_Box } from "@/components/UI/skeleton";
 import { sortDropdownData } from "@/data/uiElementsData";
+import { DEFAULT_FILTERS, SORT_DATA } from "@/features/store/productList/constants";
+import { TFilterBrands, TFilters, TListItem } from "@/features/store/productList/types";
+import { getFiltersFromProductList } from "@/features/store/productList/utils";
+import { IMAGE_BASE_URL } from "@/shared/constants/store";
 import { cn } from "@/shared/utils/styling";
-import { TListSort, TPageStatus } from "@/types/list";
-import { TBrand, TFilterBrands, TFilters, TListItem, TProductPath } from "@/types/product";
-
-const defaultFilters: TFilters = {
-  stockStatus: "all",
-  brands: [],
-  priceMinMaxLimitation: [0, 0],
-  priceMinMax: [0, 0],
-};
-
-const imgBaseUrl = process.env.IMG_URL;
-
-const sortData: TListSort[] = [
-  { sortName: "id", sortType: "desc" },
-  { sortName: "id", sortType: "asc" },
-  { sortName: "price", sortType: "desc" },
-  { sortName: "price", sortType: "asc" },
-  { sortName: "name", sortType: "asc" },
-];
+import { TPageStatus } from "@/types/list";
+import { TProductPath } from "@/types/product";
 
 const ListPage = () => {
   const router = useRouter();
@@ -46,10 +33,10 @@ const ListPage = () => {
   const [sortIndex, setSortIndex] = useState(0);
   const [showFilters, setShowFilters] = useState(false);
   const [isFilterApplied, setIsFilterApplied] = useState(false);
-  const [filters, setFilters] = useState<TFilters>(defaultFilters);
+  const [filters, setFilters] = useState<TFilters>(DEFAULT_FILTERS);
   const [appliedFilters, setAppliedFilters] = useState<TFilters>({
-    ...defaultFilters,
-    priceMinMax: [...defaultFilters.priceMinMax],
+    ...DEFAULT_FILTERS,
+    priceMinMax: [...DEFAULT_FILTERS.priceMinMax],
   });
 
   const [isListLoading, setIsListLoading] = useState(true);
@@ -58,7 +45,7 @@ const ListPage = () => {
     const getProductsList = async () => {
       setIsListLoading(true);
 
-      const response = await getList(pathName, sortData[sortIndex], appliedFilters);
+      const response = await getList(pathName, SORT_DATA[sortIndex], appliedFilters);
       if (response.error || !response.products || !response.subCategories) return router.push("/");
 
       if (isFilterApplied) {
@@ -172,7 +159,7 @@ const ListPage = () => {
         {productList.map((product) => (
           <ProductCard
             key={product.id}
-            imgUrl={[imgBaseUrl + product.images[0], imgBaseUrl + product.images[1]]}
+            imgUrl={[IMAGE_BASE_URL + product.images[0], IMAGE_BASE_URL + product.images[1]]}
             name={product.name}
             price={product.price}
             isAvailable={product.isAvailable}
@@ -290,64 +277,4 @@ const SKL_Product = (): React.ReactNode[] => {
     );
   }
   return nodes;
-};
-
-// -------- GET UNIQUE BRAND LIST FROM PRODUCT LIST --------
-const getBrandsFromProducts = (productList: TListItem[]) => {
-  return productList.map((product) => product.brand);
-};
-const removeDuplicatedBrands = (list: TBrand[]) => {
-  const newList: TBrand[] = [];
-  list.forEach((listItem) => {
-    const isFind = newList.findIndex((brand) => listItem.id === brand.id);
-    if (isFind === -1) newList.push({ id: listItem.id, name: listItem.name });
-  });
-  return newList;
-};
-const addIsSelectedValueToBrands = (brandList: TBrand[]) => {
-  return brandList.map((b) => ({
-    ...b,
-    isSelected: true,
-  }));
-};
-const generateBrands = (productList: TListItem[]) => {
-  const listOfProductsBrand: TBrand[] = getBrandsFromProducts(productList);
-  const uniqueBrandList = removeDuplicatedBrands(listOfProductsBrand);
-  return addIsSelectedValueToBrands(uniqueBrandList);
-};
-
-// -------- GET PRICE LIMIT FROM PRODUCT LIST --------
-const getPricesFromProducts = (productList: TListItem[]) => {
-  return productList.map((product) => product.price);
-};
-const findMinMax = (array: number[]) => {
-  const minMax: [number, number] = [Math.min(...array), Math.max(...array)];
-  return minMax;
-};
-const roundMaxMinPricesWithMargin = (minMax: [number, number]) => {
-  const roundedPrices: [number, number] = [...minMax];
-  roundedPrices[0] = Math.floor(roundedPrices[0]);
-  roundedPrices[0] = roundedPrices[0] - (roundedPrices[0] % 100);
-
-  roundedPrices[1] = Math.ceil(roundedPrices[1]);
-  roundedPrices[1] = roundedPrices[1] + (100 - (roundedPrices[1] % 100));
-  return roundedPrices;
-};
-const getPriceLimit = (productList: TListItem[]) => {
-  const allProductsPrices: number[] = getPricesFromProducts(productList);
-  const minMaxValues = findMinMax(allProductsPrices);
-  const roundedPrices = roundMaxMinPricesWithMargin(minMaxValues);
-
-  return roundedPrices;
-};
-
-// -------- GET INITIAL FILTERS --------
-const getFiltersFromProductList = (productsList: TListItem[]) => {
-  const newFilter: TFilters = {
-    brands: generateBrands(productsList),
-    priceMinMax: getPriceLimit(productsList),
-    priceMinMaxLimitation: getPriceLimit(productsList),
-    stockStatus: "all",
-  };
-  return newFilter;
 };
